Refactor functionality to match Database :recycle:
Refactoring of code on all pages to match updates to MongoDB
collections after first mentor call with suggestions.

With MaterializeCSS being added now for basic core
design, a lot of refactoring for all pages
and python routes were required. Working as intended and
desired.

Closed the Issue/Bug that I opened on 17APR2019 as Resolved!