Update design concept of Recipe pages
Now that I have Materialize as the framework with the basic
functionality working, I have started the design process of how
the content is displayed to the user, beginning with the recipe views.

This includes a custom toast function in javascript on page-load.

The user can now also mark ingredients and directions as complete
as they go along with the recipe, showing they have already done that
part of the recipe while following along if making it themselves.