Update styles across multiple pages :art:
I have added a couple custom styles that are applicable across
several pages, which is why a lot of them are being updated here.

Also, I have added back in the full date, not just month and year.

In preparation for filtering and sorting, I have also added
the username to be visible across all recipe cards now.