Add buttons to Print and Share recipes
Now users can Print or Share recipes directly from the recipe page.

The Print function cleans up the page without any stylings
so the user just prints the recipe name, ingredients, and directions.

The Share function populates a modal with the current URL, which
then has a button to copy it directly and advises as such.