Provide recommendations after viewing a dessert
While viewing any particular recipe, at the bottom of the
page, the user will be provided with the previous and
the next recipe in the database collection as 'recommendations'.

If the user is viewing the last recipe in the database, then
it loops over to the beginning of the collection.
If the user in viewing the first recipe in the database, then
it loops back to the last item of the collection.