Refactor CSS and a bit more JS :recycle:
I have cleaned up my CSS, added a lot of custom :root variables,
made most of the element more responsive with ems or rems
instead of px in a lot of places.

I have also fixed a few JS parsing errors as well that I
missed on the last commit of JS refactoring.